chore: add model associations
