+++ server/db/models/index.js
@@ -21,6 +21,5 @@ User.hasMany(Order)
 module.exports = {
   User,
   Wine,
-  CartItem,
   Order
 }
